refactor: improve staff management with modular components
- Add CountryCodePicker component with 35+ country dial codes
- Add DateOfBirthInput component with clear day/month/year labels
- Extract RegisterStaffModal as separate component
- Extract StaffTable as reusable table component
- Refactor StaffManagement to use new modular components
- Change from inline form to modal-based staff registration

via [Happy](https://happy.engineering)
