+++ components/settings/staff/index.ts
@@ -1,0 +1,2 @@
+export { RegisterStaffModal } from "./RegisterStaffModal";
+export { StaffTable } from "./StaffTable";
